feat: Absorb phonemizer.js and add all languages to the espeakng worker
I got tired of waiting for more languages to be added to phonemizer.js
so I just did it myself. The worker data and data chunks were
re-compiled with all available ESpeakNG languages from the original
espeakng repo and then patched in to the phonemizer.js base worker.
Keeping it here will allow for better future maintainability as well.
The licences between the projects are all the same so that should also
be a non-issue.

I've only added a test that was relevant for me and fixed up some of the
english ones with the updated worker data, but all languages should be
supported and available.

I've also gone ahead and integrated the
https://github.com/xenova/phonemizer.js/pull/4 PR, hopefully this
doesn't spring any controversy, but I'll gladly patch it away if asked.